Added Emacs configuration files.
Emacs files are built on top of grail (http://www.emacswiki.org/emacs/Grail).
The configuration takes care of the daemon mode, so that multiple frames can be
launched quickly on the same instance. Consequently, the existing emacs
configuration for mutt is removed, and so is the permanent Emacs tag for shifty.